# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Medical",
        "description": "Documentation API pour la  gestion Medical",
        "contact": {
            "email": "diopawa425@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/salles": {
            "get": {
                "tags": [
                    "salles"
                ],
                "summary": "liste des salles",
                "description": "Liste de tous les salles.",
                "operationId": "listSalles",
                "responses": {
                    "200": {
                        "description": "Données récupérées avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé, token invalide ou manquant."
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "salles"
                ],
                "summary": "Créer un nouveau salle",
                "description": "Enregistre un nouveau salle dans la base de données.",
                "operationId": "createSalle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Salle"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Salle"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/salle/service/{id}": {
            "get": {
                "tags": [
                    "salles"
                ],
                "summary": "Get salles by service",
                "description": "Afficher les infos des salles d'une service.",
                "operationId": "GetSaleesByService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la service",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personnel trouvé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Salles trouvé avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Salle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personnel non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Salles non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/salles/{id}": {
            "get": {
                "tags": [
                    "salles"
                ],
                "summary": "Get One by Id",
                "description": "Afficher les infos d'une salle admission.",
                "operationId": "GetOneSalleAdmission",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle à afficher",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salle trouvé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Salle trouvé avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Salle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Salle non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Salle non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "salles"
                ],
                "summary": "Modifier les infos d'une salle",
                "description": "Modifier les infos d'une salle.",
                "operationId": "updateSalles",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Salle"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Salle mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Salle mis à jour avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Salle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Salle non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Salle non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Salle": {
                "title": "Salle",
                "description": "Gestion des salles",
                "properties": {
                    "service_id": {
                        "description": "Service associée",
                        "type": "integer",
                        "example": "1"
                    },
                    "nom": {
                        "description": "Nom de la salle",
                        "type": "string",
                        "example": "A1"
                    },
                    "capacite": {
                        "description": "Nombre de place",
                        "type": "string",
                        "example": "5"
                    },
                    "type": {
                        "description": "Type de salle",
                        "type": "string",
                        "example": "chambre"
                    },
                    "disponible": {
                        "description": "Disponibilite de la salle",
                        "type": "boolean",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Entrez le token Sanctum reçu après connexion.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}